+++ internal/exceptions/Exceptions.go
